styles folder and move css files
